+++ src/main/java/com/harium/etyl/animation/script/text/DialogScript.java
@@ -5,24 +5,22 @@ import com.harium.etyl.layer.TextLayer;
 
 public class DialogScript extends AnimationScript {
 
-	private String fullText;
-	
-	private TextLayer target;
-	
-	public DialogScript(TextLayer target, long time) {
-		super(time);
-		
-		this.target = target;
-		
-		fullText = target.getText();
-	}
-	
-	@Override
-	public void calculate(double factor) {
-		
-		int lastChar = (int)(fullText.length()*factor);
-		
-		target.setText(fullText.substring(0, lastChar));
-	}
-	
+    private String fullText;
+
+    private TextLayer target;
+
+    public DialogScript(TextLayer target, long time) {
+        super(time);
+
+        this.target = target;
+
+        fullText = target.getText();
+    }
+
+    @Override
+    public void calculate(float factor) {
+        int lastChar = (int) (fullText.length() * factor);
+        target.setText(fullText.substring(0, lastChar));
+    }
+
 }
